delete order by user

+++ order-fastapi/routers/orders.py
@@ -66,10 +66,12 @@ async def update_order_by_id(user: user_dependency, db: Annotated[Session, Depen
 
 
 @routers.delete('/orders/{order_id}')
-async def delete_order_by_id(db: Annotated[Session, Depends(get_db)], order_id: int):
-    order_model = db.query(Orders).filter(Orders.id == order_id).first()
+async def delete_order_by_id(user: user_dependency, db: Annotated[Session, Depends(get_db)], order_id: int):
+    if user is None:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="Authentication is failed!")
+    order_model = db.query(Orders).filter(Orders.id == order_id).filter(Orders.owner_id == user.get('id')).first()
     if order_model is None:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Order not found")
-    db.query(Orders).filter(Orders.id == order_id).delete()
+    db.query(Orders).filter(Orders.id == order_id).filter(Orders.owner_id == user.get('id')).delete()
     db.commit()
